pyproject.toml
[project]
name = "agent-level-vs-run-level-vs-global-level-configuration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
]
